+++ bindings/python/ContinuousDynamicalSystem/src/Module.cpp
@@ -7,12 +7,13 @@
 
 #include <pybind11/pybind11.h>
 
+#include <BipedalLocomotion/bindings/ContinuousDynamicalSystem/ButterworthLowPassFilter.h>
 #include <BipedalLocomotion/bindings/ContinuousDynamicalSystem/CentroidalDynamics.h>
-#include <BipedalLocomotion/bindings/ContinuousDynamicalSystem/FloatingBaseSystemKinematics.h>
+#include <BipedalLocomotion/bindings/ContinuousDynamicalSystem/FloatingBaseSystemAccelerationKinematics.h>
+#include <BipedalLocomotion/bindings/ContinuousDynamicalSystem/FloatingBaseSystemVelocityKinematics.h>
 #include <BipedalLocomotion/bindings/ContinuousDynamicalSystem/LinearTimeInvariantSystem.h>
 #include <BipedalLocomotion/bindings/ContinuousDynamicalSystem/Module.h>
 #include <BipedalLocomotion/bindings/ContinuousDynamicalSystem/MultiStateWeightProvider.h>
-#include <BipedalLocomotion/bindings/ContinuousDynamicalSystem/ButterworthLowPassFilter.h>
 
 namespace BipedalLocomotion
 {
@@ -26,7 +27,8 @@ void CreateModule(pybind11::module& module)
                    "BipedalLocomotion::ContinuousDynamicalSystem";
 
     CreateLinearTimeInvariantSystem(module);
-    CreateFloatingBaseSystemKinematics(module);
+    CreateFloatingBaseSystemVelocityKinematics(module);
+    CreateFloatingBaseSystemAccelerationKinematics(module);
     CreateMultiStateWeightProvider(module);
     CreateCentroidalDynamics(module);
     CreateButterworthLowPassFilter(module);